requirements.txt
git+https://github.com/modelcontextprotocol/python-sdk.git@main
fastapi
uvicorn
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-monitor
